Sync

+++ indexer/httpd/src/graphql/types/block.rs
@@ -10,7 +10,7 @@ use {
 #[graphql(complex)]
 #[serde(default)]
 pub struct Block {
-    pub block_height: i64,
+    pub block_height: u64,
     pub created_at: DateTime<Utc>,
     pub hash: String,
     pub app_hash: String,
@@ -19,7 +19,7 @@ pub struct Block {
 impl From<entity::blocks::Model> for Block {
     fn from(item: entity::blocks::Model) -> Self {
         Self {
-            block_height: item.block_height,
+            block_height: item.block_height as u64,
             created_at: Utc.from_utc_datetime(&item.created_at),
             hash: item.hash,
             app_hash: item.app_hash,
